pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "purview-migration-toolkit"
version = "0.1.0"
description = "Export and import Microsoft Purview artifacts across tenants."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "azure-identity>=1.17.1",
  "requests>=2.32.3",
]

[project.scripts]
purview-migrate = "purview_migration.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
